+++ lib/claper/repo.ex
@@ -4,6 +4,10 @@ defmodule Claper.Repo do
     adapter: Ecto.Adapters.Postgres
 
   def init(_type, config) do
-    {:ok, Keyword.put(config, :url, System.get_env("DATABASE_URL"))}
+    if url = System.get_env("DATABASE_URL") do
+      {:ok, Keyword.put(config, :url, url)}
+    else
+      {:ok, config}
+    end
   end
 end